requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
asyncpg==0.29.0
passlib[argon2]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
pydantic==2.8.2
email-validator==2.2.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.8